pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agentflow"
version = "1.0.0"
description = "Multi-Agent AI System – Gen AI Academy APAC Hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic-settings>=2.0.0",
    "google-cloud-aiplatform>=1.43.0",
    "langchain-google-vertexai>=0.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.5",
]
license = { text = "MIT" }
keywords = ["ai", "agents", "gemini", "adk", "mcp", "fastapi"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
log_cli = true
log_cli_level = "INFO"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

[tool.coverage.run]
source = ["backend"]
omit = ["*/__pycache__/*", "*/tests/*", "*/.venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.flake8]
max-line-length = 120
ignore = ["E501", "W503", "E203"]
exclude = [".venv", "__pycache__", "build", "dist", "alembic/versions"]
per-file-ignores = [
    "backend/database/models.py:E501",
    "tests/*:E501",
]
